Update search index correctly.

+++ Sources/SplitTopics.php
@@ -1341,14 +1341,18 @@ function MergeExecute($topics = array())
 	$updated_topics = array();
 
 	// Create stub topics out of the remaining topics.
-	// We don't want the search index data though.
-	$smcFunc['db_query']('', '
+	// We don't want the search index data though (For non-redirect merges).
+    if (!isset($_POST['postRedirect']))
+    {
+        $smcFunc['db_query']('', '
 		DELETE FROM {db_prefix}log_search_subjects
 		WHERE id_topic IN ({array_int:deleted_topics})',
-		array(
-			'deleted_topics' => $deleted_topics,
-		)
-	);
+            array(
+                'deleted_topics' => $deleted_topics,
+            )
+        );
+    }
+
 	require_once($sourcedir . '/Subs-Post.php');
 	$posterOptions = array(
 		'id' => $user_info['id'],
@@ -1374,9 +1378,11 @@ function MergeExecute($topics = array())
 
 		foreach ($deleted_topics as $this_old_topic)
 		{
+            $redirect_subject = sprintf($txt['merged_subject'], $topic_data[$this_old_topic]['subject']);
+
 			$msgOptions = array(
 				'icon' => 'moved',
-				'subject' => sprintf($txt['merged_subject'], $topic_data[$this_old_topic]['subject']),
+				'subject' => $redirect_subject,
 				'body' => $reason,
 				'approved' => 1,
 			);
@@ -1394,6 +1400,9 @@ function MergeExecute($topics = array())
 			{
 				$updated_topics[$this_old_topic] = $msgOptions['id'];
 			}
+
+            // Update subject search index
+            updateStats('subject',$this_old_topic,$redirect_subject);
 		}
 	}
